# storage/api-docs/api-docs.json
{"swagger":"2.0","info":{"title":"API VayeCongress","version":"1.0.0"},"basePath":"\/congress-backend-modules\/public\/api","paths":{"\/access\/type\/list":{"get":{"summary":"Get All Type Of Access","operationId":"getAllTypesAccess","responses":{"200":{"description":"successful operation"},"406":{"description":"not acceptable"},"500":{"description":"internal server error"}},"security":[{"Bearer":[]}]}},"\/add-info\/type\/list":{"get":{"summary":"Get All Type Info","operationId":"getAllTypesInfo","responses":{"200":{"description":"successful operation"},"406":{"description":"not acceptable"},"500":{"description":"internal server error"}},"security":[{"Bearer":[]}]}},"\/mobile\/scan\/participant":{"post":{"tags":["Mobile"],"summary":"Scan Participant","operationId":"scanParticipatorQrCode","parameters":[{"name":"QrCode","in":"query","description":"QR code","required":true,"type":"string"}],"responses":{"200":{"description":"successful operation"},"406":{"description":"not acceptable"},"500":{"description":"internal server error"}},"security":[{"Bearer":[]}]}},"\/mobile\/presence\/{participantId}\/status\/update":{"post":{"tags":["Mobile"],"summary":"makeUserPresentCongress","operationId":"makeUserPresentCongress","parameters":[{"name":"participantId","in":"path","description":"participantId","required":true,"type":"integer"},{"name":"isPresent","in":"query","description":"isPresent","required":true,"type":"integer"},{"name":"congressId","in":"query","description":"L'id du congr\u00e8s","required":true,"type":"integer"}],"responses":{"200":{"description":"successful operation"},"406":{"description":"not acceptable"},"500":{"description":"internal server error"}},"security":[{"Bearer":[]}]}},"\/mobile\/presence\/{participantId}\/status\/update\/access":{"post":{"tags":["Mobile"],"summary":"makeUserPresentAccess","operationId":"makeUserPresentAccess","parameters":[{"name":"participantId","in":"path","description":"participantId","required":true,"type":"integer"},{"name":"isPresent","in":"query","description":"isPresent","required":true,"type":"integer"},{"name":"accessId","in":"query","description":"L'id de l'acc\u00e9es","required":true,"type":"integer"}],"responses":{"200":{"description":"successful operation"},"406":{"description":"not acceptable"},"500":{"description":"internal server error"}},"security":[{"Bearer":[]}]}},"\/admin\/me":{"get":{"summary":"Get Admin By Token","operationId":"getAuhentificatedAdmin","responses":{"200":{"description":"successful operation"},"406":{"description":"not acceptable"},"500":{"description":"internal server error"}},"security":[{"Bearer":[]}]}},"\/admin\/me\/congress":{"get":{"summary":"Get Congress By Admin","operationId":"getAdminCongresses","responses":{"200":{"description":"successful operation"},"406":{"description":"not acceptable"},"500":{"description":"internal server error"}},"security":[{"Bearer":[]}]}},"\/admin\/me\/personels\/list":{"get":{"summary":"Get personels by Admin","operationId":"getListPersonels","responses":{"200":{"description":"successful operation"},"406":{"description":"not acceptable"},"500":{"description":"internal server error"}},"security":[{"Bearer":[]}]}},"\/auth\/login\/admin":{"post":{"summary":"Login Admin","operationId":"loginAdmin","parameters":[{"name":"email","in":"query","description":"Email Admin.","required":true,"type":"string"},{"name":"password","in":"query","description":"Password Admin.","required":true,"type":"string"}],"responses":{"200":{"description":"successful operation"},"406":{"description":"not acceptable"},"500":{"description":"internal server error"}}}},"\/mobile\/login":{"post":{"tags":["Mobile"],"summary":"Login","operationId":"loginAdminMobile","parameters":[{"name":"QrCode","in":"query","description":"QR code","required":true,"type":"string"}],"responses":{"200":{"description":"successful operation"},"406":{"description":"not acceptable"},"500":{"description":"internal server error"}}}},"\/mobile\/congress":{"get":{"tags":["Mobile"],"summary":"getCongressByAdmin","operationId":"getCongressByAdmin","responses":{"200":{"description":"successful operation"},"406":{"description":"not acceptable"},"500":{"description":"internal server error"}},"security":[{"Bearer":[]}]}}},"definitions":{},"securityDefinitions":{"Bearer":{"type":"apiKey","name":"Authorization","in":"header"}},"components":{"securitySchemes":{"jwt":{"type":"apiKey","description":"A short description for security scheme","name":"Authorization","in":"header"}}}}